# openapi/contabilidad-api.yaml
openapi: 3.1.0
info:
  title: Contabilidad API
  description: |
    API REST para el sistema de contabilidad y facturación electrónica.
    Soporta CFDI 4.0, REP 2.0, catálogos SAT versionados, motor fiscal configurable
    y preparación para contabilidad electrónica.
  version: 0.1.0
  contact:
    name: Contabilidad Team
servers:
  - url: http://localhost:8080
    description: Desarrollo local
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sesión iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Renovar token de acceso
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token renovado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Cerrar sesión
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: Sesión cerrada
...
  /api/v1/companies:
    post:
      tags: [Companies]
      summary: Crear empresa
      operationId: createCompany
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Empresa creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDto'
  /api/v1/companies/{companyId}:
    get:
      tags: [Companies]
      summary: Obtener empresa
      operationId: getCompany
      parameters:
        - $ref: '#/components/parameters/companyId'
      responses:
        '200':
          description: Empresa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDto'
    patch:
      tags: [Companies]
      summary: Actualizar empresa
      operationId: updateCompany
      parameters:
        - $ref: '#/components/parameters/companyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyRequest'
      responses:
        '200':
          description: Empresa actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyDto'

  # ── Clients ──
...
  /api/v1/invoices:
    get:
      tags: [Invoices]
      summary: Listar facturas
      operationId: listInvoices
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, VALIDATED, STAMPED, CANCELLED]
      responses:
        '200':
          description: Lista de facturas
  /api/v1/invoices/drafts:
    post:
      tags: [Invoices]
      summary: Crear borrador de factura
      operationId: createInvoiceDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceDraftRequest'
      responses:
        '201':
          description: Borrador creado
  /api/v1/invoices/{invoiceId}:
    get:
      tags: [Invoices]
      summary: Obtener factura
      operationId: getInvoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Factura
  /api/v1/invoices/{invoiceId}/validate:
    post:
      tags: [Invoices]
      summary: Validar factura antes de timbrar
      operationId: validateInvoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Resultado de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
  /api/v1/invoices/{invoiceId}/stamp:
    post:
      tags: [Invoices]
      summary: Timbrar CFDI
      operationId: stampInvoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Timbrado aceptado para proceso asíncrono
        '409':
          description: Documento ya timbrado o petición duplicada
        '422':
          description: Documento no fiscalmente válido
  /api/v1/invoices/{invoiceId}/cancel:
    post:
      tags: [Invoices]
      summary: Cancelar CFDI
      operationId: cancelInvoice
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: CFDI cancelado
        '422':
          description: No se puede cancelar
...
  /api/v1/expenses:
    get:
      tags: [Expenses]
      summary: Listar gastos
      operationId: listExpenses
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Lista de gastos
    post:
      tags: [Expenses]
      summary: Crear gasto manual
      operationId: createExpense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseRequest'
      responses:
        '201':
          description: Gasto creado
  /api/v1/expenses/import-xml:
    post:
      tags: [Expenses]
      summary: Importar gasto desde XML CFDI
      operationId: importExpenseXml
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Gasto importado desde XML
  /api/v1/expenses/{expenseId}:
    get:
      tags: [Expenses]
      summary: Obtener gasto
      operationId: getExpense
      parameters:
        - name: expenseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Gasto

  # ── Payments ──
...
  /api/v1/declarations:
    get:
      tags: [Declarations]
      summary: Listar declaraciones
      operationId: listDeclarations
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
        - name: year
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [MONTHLY, ANNUAL, DIOT, INFORMATIVE]
      responses:
        '200':
          description: Lista de declaraciones
  /api/v1/declarations/monthly-workpapers:generate:
    post:
      tags: [Declarations]
      summary: Generar papeles de trabajo mensuales
      operationId: generateMonthlyWorkpapers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateWorkpapersRequest'
      responses:
        '201':
          description: Papeles generados
  /api/v1/declarations/annual-summary:generate:
    post:
      tags: [Declarations]
      summary: Generar resumen anual
      operationId: generateAnnualSummary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [year]
              properties:
                year:
                  type: integer
      responses:
        '201':
          description: Resumen anual generado
  /api/v1/declarations/{declarationId}:
    get:
      tags: [Declarations]
      summary: Obtener declaración
      operationId: getDeclaration
      parameters:
        - name: declarationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Declaración

  # ── Audit ──
  /api/v1/audit:
    get:
      tags: [Audit]
      summary: Listar bitácora de auditoría
      operationId: listAuditLog
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/size'
      responses:
        '200':
          description: Bitácora
  /api/v1/audit/entity/{entityType}/{entityId}:
    get:
      tags: [Audit]
      summary: Auditoría por entidad
      operationId: getEntityAudit
      parameters:
        - name: entityType
          in: path
          required: true
          schema:
            type: string
        - name: entityId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Registros de auditoría
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    companyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    size:
      name: size
      in: query
      schema:
        type: integer
        default: 10
...
  schemas:
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Segundos hasta expiración
...
    CompanyDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        rfc:
          type: string
        legalName:
          type: string
        taxpayerType:
          type: string
          enum: [PERSONA_MORAL, PERSONA_FISICA]
        fiscalRegimeCode:
          type: string
        taxZoneProfile:
          type: string
        postalCode:
          type: string
        active:
          type: boolean
    CreateCompanyRequest:
      type: object
      required: [rfc, legalName, taxpayerType, fiscalRegimeCode]
      properties:
        rfc:
          type: string
          minLength: 12
          maxLength: 13
        legalName:
          type: string
        taxpayerType:
          type: string
          enum: [PERSONA_MORAL, PERSONA_FISICA]
        fiscalRegimeCode:
          type: string
        postalCode:
          type: string
    UpdateCompanyRequest:
      type: object
      properties:
        legalName:
          type: string
        fiscalRegimeCode:
          type: string
        taxZoneProfile:
          type: string
        postalCode:
          type: string

    # ── Clients ──
...
    CreateInvoiceDraftRequest:
      type: object
      required: [clientId, invoiceType, paymentMethodCode, usoCfdiCode]
      properties:
        clientId:
          type: string
          format: uuid
        invoiceType:
          type: string
          enum: [I, E, T, P]
        paymentMethodCode:
          type: string
          enum: [PUE, PPD]
        paymentFormCode:
          type: string
        usoCfdiCode:
          type: string
        currencyCode:
          type: string
          default: MXN
        lines:
          type: array
          items:
            $ref: '#/components/schemas/InvoiceLineRequest'
    InvoiceLineRequest:
      type: object
      required: [description, quantity, unitPrice, satProductCode, satUnitCode]
      properties:
        productId:
          type: string
          format: uuid
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        satProductCode:
          type: string
        satUnitCode:
          type: string
        taxProfileId:
          type: string
          format: uuid
    ValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            type: string
    CancelRequest:
      type: object
      required: [reasonCode]
      properties:
        reasonCode:
          type: string
          enum: ['01', '02', '03', '04']
        replacementUuid:
          type: string
          format: uuid
          description: Requerido cuando reasonCode es '01'

    # ── Expenses ──
    CreateExpenseRequest:
      type: object
      required: [supplierId, total]
      properties:
        supplierId:
          type: string
          format: uuid
        total:
          type: number
        subtotal:
          type: number
        currencyCode:
          type: string
          default: MXN
        category:
          type: string
        notes:
          type: string

    # ── Payments ──
...
    GenerateWorkpapersRequest:
      type: object
      required: [year, month]
      properties:
        year:
          type: integer
        month:
          type: integer
          minimum: 1
          maximum: 12

    # ── Common ──
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              rejectedValue: {}